Refactor : Post 엔티티로 필드 위임
